backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
httpx
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
